Consolidate Google token resolution into one credential broker
Three copies of the same better-auth getAccessToken({providerId:"google"})
logic had drifted apart: the shared getGoogleAccessToken in lib/calendarOps,
a private duplicate in googleSync (typed GenericCtx so it runs from crons), and
a third inlined inside acceptBooking. An assistant-driven accept even resolved
the token twice.

Extract one broker (lib/googleCredentials.getGoogleAccessToken) typed on the
broadest GenericCtx<DataModel> every caller satisfies, and route calendar
writes, background sync, and booking acceptance through it. This is the single
choke point for provider credentials the connection model will later make
connection-aware; today connection == the user's Google login grant.

No behavior change. Backend 73 unit + 19 integration + types green.
